refactor: extract Server, Monitor, Backup services from bridge.go (Phase 3)
Agent-Task: bridge-refactor-phase3
Agent-Model: SOLO
Agent-Decision: Extract 3 domain service files to reduce bridge.go from 2094 to 1735 lines

Extracted services:
- server_service.go (189 lines): ListServers, AddServer, RemoveServer, TestServer,
  UpdateServer, GetRemoteExecutorForTerminal, GetServersByTags
- monitor_service.go (52 lines): GetContainerMetrics, GetSystemMetrics, ListAlerts,
  ListAlertRules, getMonitor, getHealer
- backup_service.go (156 lines): shellQuote, Backup, Restore, BatchBackup

All methods remain on Bridge struct receiver to satisfy mcp.Deployer interface.
Removed unused filepath import from bridge.go.
